Refactored dataset-related classes from `org.cirdles.topsoil.app.utils`
Refactored dataset-related classes from `org.cirdles.topsoil.app.utils`
into `org.cirdles.topsoil.app.dataset` with appropriate subpackages for
dataset readers and writers.

Also introduced the abstract class `BaseDatasetReader` and removed
default methods from `DatasetReader`.
