+++ src/main/java/org/cirdles/topsoil/app/TSVTable.java
@@ -26,10 +26,10 @@ import javafx.scene.input.Clipboard;
 import javafx.scene.input.KeyCode;
 import javafx.scene.input.KeyEvent;
 import org.cirdles.topsoil.app.table.EntryTableColumn;
-import org.cirdles.topsoil.app.utils.TSVDatasetReader;
-import org.cirdles.topsoil.app.utils.TSVDatasetWriter;
-import org.cirdles.topsoil.app.utils.DatasetReader;
-import org.cirdles.topsoil.app.utils.DatasetWriter;
+import org.cirdles.topsoil.app.dataset.reader.TSVDatasetReader;
+import org.cirdles.topsoil.app.dataset.writer.TSVDatasetWriter;
+import org.cirdles.topsoil.app.dataset.reader.DatasetReader;
+import org.cirdles.topsoil.app.dataset.writer.DatasetWriter;
 import org.cirdles.topsoil.dataset.Dataset;
 import org.cirdles.topsoil.dataset.entry.Entry;
 import org.cirdles.topsoil.dataset.field.NumberField;
