+++ src/main/java/org/cirdles/topsoil/app/Tools.java
@@ -28,10 +28,10 @@ import javafx.scene.control.TableView;
 import javafx.scene.input.Clipboard;
 import javafx.util.StringConverter;
 import static org.cirdles.topsoil.app.Topsoil.LAST_TABLE_PATH;
-import org.cirdles.topsoil.app.utils.TSVDatasetReader;
-import org.cirdles.topsoil.app.utils.TSVDatasetWriter;
-import org.cirdles.topsoil.app.utils.DatasetReader;
-import org.cirdles.topsoil.app.utils.DatasetWriter;
+import org.cirdles.topsoil.app.dataset.reader.TSVDatasetReader;
+import org.cirdles.topsoil.app.dataset.writer.TSVDatasetWriter;
+import org.cirdles.topsoil.app.dataset.reader.DatasetReader;
+import org.cirdles.topsoil.app.dataset.writer.DatasetWriter;
 import org.cirdles.topsoil.app.utils.YesNoAlert;
 import org.cirdles.topsoil.dataset.Dataset;
 import org.cirdles.topsoil.dataset.entry.Entry;
