+++ src/main/java/org/cirdles/topsoil/app/TopsoilMainWindow.java
@@ -42,12 +42,12 @@ import javafx.stage.Window;
 import org.cirdles.javafx.CustomVBox;
 import org.cirdles.topsoil.app.chart.ChartWindow;
 import org.cirdles.topsoil.app.chart.VariableBindingDialog;
-import org.cirdles.topsoil.app.utils.DatasetReader;
-import org.cirdles.topsoil.app.utils.DatasetWriter;
+import org.cirdles.topsoil.app.dataset.reader.DatasetReader;
+import org.cirdles.topsoil.app.dataset.writer.DatasetWriter;
 import org.cirdles.topsoil.app.utils.GetApplicationDirectoryOperation;
-import org.cirdles.topsoil.app.utils.TSVDatasetManager;
-import org.cirdles.topsoil.app.utils.TSVDatasetReader;
-import org.cirdles.topsoil.app.utils.TSVDatasetWriter;
+import org.cirdles.topsoil.app.dataset.TSVDatasetManager;
+import org.cirdles.topsoil.app.dataset.reader.TSVDatasetReader;
+import org.cirdles.topsoil.app.dataset.writer.TSVDatasetWriter;
 import org.cirdles.topsoil.chart.Chart;
 import org.cirdles.topsoil.chart.JavaScriptChart;
 import org.cirdles.topsoil.dataset.Dataset;
