+++ src/main/java/org/cirdles/topsoil/app/dataset/reader/DSVDatasetReader.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.cirdles.topsoil.app.utils;
+package org.cirdles.topsoil.app.dataset.reader;
 
 import au.com.bytecode.opencsv.CSVReader;
 import java.io.IOException;
@@ -35,7 +35,7 @@ import org.cirdles.topsoil.dataset.field.TextField;
  *
  * @author John Zeringue
  */
-public class DSVDatasetReader implements DatasetReader {
+public class DSVDatasetReader extends BaseDatasetReader {
     
     private char delimiter;
     private boolean expectingHeaders;
