+++ src/main/java/org/cirdles/topsoil/app/dataset/reader/TSVDatasetReader.java
@@ -13,16 +13,20 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.cirdles.topsoil.app.utils;
+package org.cirdles.topsoil.app.dataset.reader;
 
 /**
  *
  * @author John Zeringue <john.joseph.zeringue@gmail.com>
  */
-public class TSVDatasetWriter extends DSVDatasetWriter {
+public class TSVDatasetReader extends DSVDatasetReader {
 
-    public TSVDatasetWriter() {
+    public TSVDatasetReader() {
         super('\t');
+    }
+
+    public TSVDatasetReader(boolean expectingHeaders) {
+        super('\t', expectingHeaders);
     }
 
 }
