+++ src/main/java/org/cirdles/topsoil/app/dataset/writer/DSVDatasetWriter.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.cirdles.topsoil.app.utils;
+package org.cirdles.topsoil.app.dataset.writer;
 
 import au.com.bytecode.opencsv.CSVWriter;
 import java.io.IOException;
