+++ src/main/java/org/cirdles/topsoil/data/entry/Entry.java
