+++ src/test/java/org/cirdles/topsoil/app/table/EntryTableColumnTest.java
@@ -25,10 +25,10 @@ import javafx.stage.Stage;
 import org.cirdles.topsoil.app.TSVTable;
 import org.cirdles.topsoil.app.TSVTableTest;
 import org.cirdles.topsoil.app.Topsoil;
-import org.cirdles.topsoil.app.utils.DatasetReader;
-import org.cirdles.topsoil.app.utils.DatasetWriter;
-import org.cirdles.topsoil.app.utils.TSVDatasetReader;
-import org.cirdles.topsoil.app.utils.TSVDatasetWriter;
+import org.cirdles.topsoil.app.dataset.reader.DatasetReader;
+import org.cirdles.topsoil.app.dataset.writer.DatasetWriter;
+import org.cirdles.topsoil.app.dataset.reader.TSVDatasetReader;
+import org.cirdles.topsoil.app.dataset.writer.TSVDatasetWriter;
 import org.cirdles.topsoil.dataset.Dataset;
 import org.junit.Test;
 import static org.testfx.api.FxAssert.verifyThat;
